# packages/contracts-openapi/dist/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Aistroyka API",
    "version": "1.0.0",
    "description": "API v1 for Aistroyka Construction Intelligence. Auth: Bearer JWT + tenant context. Mobile: send x-device-id on sync and x-idempotency-key on all writes. Errors: 400/401/403/404/409/413/429 with JSON { error?: string }."
  },
  "servers": [
    {
      "url": "https://api.aistroyka.ai",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/api/v1/ai/analyze-image": {
      "post": {
        "summary": "Analyze construction image",
        "tags": [
          "AI"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeImageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Analysis result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResult"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeImageError"
                }
              }
            }
          },
          "402": {
            "description": "Quota exceeded"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalyzeImageRequest": {
        "$ref": "#/definitions/AnalyzeImageRequest",
        "definitions": {
          "AnalyzeImageRequest": {
            "type": "object",
            "properties": {
              "image_url": {
                "type": "string"
              },
              "media_id": {
                "type": "string"
              },
              "project_id": {
                "type": "string"
              }
            },
            "required": [
              "image_url"
            ],
            "additionalProperties": false
          }
        }
      },
      "AnalysisResult": {
        "$ref": "#/definitions/AnalysisResult",
        "definitions": {
          "AnalysisResult": {
            "type": "object",
            "properties": {
              "stage": {
                "type": "string"
              },
              "completion_percent": {
                "type": "number"
              },
              "risk_level": {
                "type": "string",
                "enum": [
                  "low",
                  "medium",
                  "high"
                ]
              },
              "detected_issues": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "recommendations": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "required": [
              "stage",
              "completion_percent",
              "risk_level",
              "detected_issues",
              "recommendations"
            ],
            "additionalProperties": false
          }
        }
      },
      "AnalyzeImageError": {
        "$ref": "#/definitions/AnalyzeImageError",
        "definitions": {
          "AnalyzeImageError": {
            "type": "object",
            "properties": {
              "error": {
                "type": "string"
              }
            },
            "required": [
              "error"
            ],
            "additionalProperties": false
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}